aws_telegram_bot/requirements.txt
appnope==0.1.0
awscli==1.10.13
boto3==1.3.0
botocore==1.4.4
colorama==0.3.3
decorator==4.0.9
docutils==0.12
future==0.15.2
futures==3.0.5
#gnureadline==6.3.3
ipython==4.1.2
ipython-genutils==0.1.0
jmespath==0.9.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
pyasn1==0.1.9
python-dateutil==2.5.0
python-telegram-bot==3.3
pytz==2016.1
rsa==3.3
s3transfer==0.0.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
